delete items from custom lists

+++ app.js
@@ -8,9 +8,6 @@ const date = require(__dirname + "/date.js"); // ALWAYS REMEMBER ////// SLASH
 const app = express();
 
 //can use const cause we can push to the array but not assign new array
-
-const items = ["Buy groceries", "Go to post office", "Pay bills"];
-const workItems = [];
 const day = date.getDate(); //can change to date.getDay()
 
 app.set("view engine", "ejs"); //below the express one cause uses app
@@ -20,7 +17,8 @@ app.use(express.static("public"));
 
 mongoose.connect("mongodb://localhost:27017/todolistDB", {
   useNewUrlParser: true,
-  useUnifiedTopology: true
+  useUnifiedTopology: true,
+  useFindAndModify: false
 });
 //create
 const itemSchema = {
@@ -45,6 +43,14 @@ const listSchema = {
   items: [itemSchema]
 };
 const List = mongoose.model("List", listSchema);
+
+app.get("/about", (req, res) => {
+  res.render("about");
+});
+
+app.get("/favicon.ico", function(req, res) {
+  res.sendStatus(204);
+});
 
 app.get("/", (req, res) => {
   Item.find({}, (err, foundItems) => {
@@ -85,20 +91,32 @@ app.post("/", (req, res) => {
     });
   }
 });
+
 app.post("/delete", (req, res) => {
   const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, err => {
-    if (!err) {
-      console.log("Successfully removed.");
-      res.redirect("/");
-    }
-  });
+  if ((listName === day)) {
+    Item.findByIdAndRemove(checkedItemId, err => {
+      if (!err) {
+        console.log("Successfully removed.");
+        res.redirect("/");
+      }
+    });
+  } else {
+    console.log(listName, checkedItemId);
+    List.findOneAndUpdate(
+      { name: listName },
+      { $pull: { items: { _id: checkedItemId } } },
+      (err, foundList) => {
+        if (!err) {
+          res.redirect("/" + listName);
+        }
+      }
+    );
+  }
 });
 
-app.get("/favicon.ico", function(req, res) {
-  res.sendStatus(204);
-});
 
 app.get("/:customListName", (req, res) => {
   const customListName = req.params.customListName;
@@ -122,9 +140,6 @@ app.get("/:customListName", (req, res) => {
   });
 });
 
-app.get("/about", (req, res) => {
-  res.render("about");
-});
 
 app.listen(process.env.PORT || 3000, () => {
   console.log("Server started on port 3000.");
